chatbot-server/requirements.txt
fastapi[all]
uvicorn==0.34.0
pydantic==2.10.6
requests>=2.31.0
python-dotenv==1.0.1
mistralai>=1.0.0
langchain-core>=0.1.0
PyMuPDF==1.26.0
arxiv==2.1.0
chromadb==0.4.24
sentence-transformers>=2.5.1
transformers>=4.35.0
torch>=2.1.1
torchvision>=0.15.0
numpy>=1.24.0